# swagger.yml
openapi: 3.0.0
info:
  title: Cloud Evaluator API
  description: API documentation for the Cloud Evaluator platform
  version: 1.0.0
servers:
  - url: /api
    description: API base path updated to start with /api
paths:
...
  /proctor/start:
    post:
      summary: Start a proctoring session
      tags:
        - Proctoring
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                studentAssessmentId:
                  type: integer
                  example: 1
                proctorId:
                  type: integer
                  example: 2
      responses:
        '201':
          description: Proctoring session started successfully
        '400':
          description: Missing required fields
        '500':
          description: Failed to start proctoring session
  /proctor/violation:
    post:
      summary: Record a violation during a proctoring session
      tags:
        - Proctoring
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: integer
                  example: 1
                violationType:
                  type: string
                  example: "cheating"
                evidence:
                  type: object
                  example: { screenshot: "base64string" }
      responses:
        '200':
          description: Violation recorded successfully
        '400':
          description: Missing required fields
        '500':
          description: Failed to record violation
  /proctor/recording:
    post:
      summary: Save a recording for a proctoring session
      tags:
        - Proctoring
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: integer
                  example: 1
                recordingType:
                  type: string
                  enum: ["screen", "camera"]
                  example: "screen"
                filePath:
                  type: string
                  example: "/recordings/session1/screen.mp4"
      responses:
        '200':
          description: Recording saved successfully
        '400':
          description: Missing required fields
        '500':
          description: Failed to save recording
  /auth/login:
    post:
      summary: User login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "user@example.com"
                password:
                  type: string
                  example: "password123"
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT